+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.4] - 2018-07-12
+### Added
+- [PHP] Add AuthorizeByTag
 ### Fixed
-- Fixed PaginationHelper not to filter 0 value of query string #54
+- [PHP] Fixed PaginationHelper not to filter 0 value of query string #54
 ### Changed
+- [PHP] Use new authorize endpoint `/auth/oauth2/authorize`
 - [PHP] Use XDEBUG_ENABLE instead DEBUG env for XDebug switch
 
 ## [2.3.3] - 2018-05-24